Hero and Element subclassing
Simplified the character class and created a Hero subclass with
specific statistical distinctions like gold and experience. Also
changed element to be a member variable of the character class and so
each element is a subclass of the parent Element class. Also expanded
the game to change the map size depending on the difficulty of the
game, as well as increasing and decreasing the strength of monsters in
the game according to difficulty. Finally made the Character name field
private and now must access it with an accessor method.